+++ R/plot_tt.R
@@ -40,14 +40,9 @@ plot_tt <- function(x,
 
   out <- x
 
-  # j is a regular expression
-  # before assertions
-  if (is.character(j) && !is.null(meta(x, "colnames"))) {
-    j <- grep(j, meta(x, "colnames"), perl = TRUE)
-  }
 
+  j <- sanitize_j(j, x)
   assert_integerish(i, null.ok = TRUE)
-  assert_integerish(j, null.ok = TRUE)
   assert_numeric(height, len = 1, lower = 0)
   assert_numeric(asp, len = 1, lower = 0, upper = 1)
   assert_class(x, "tinytable")
@@ -74,7 +69,7 @@ plot_tt <- function(x,
   }
 
   if (is.character(fun)) {
-    assert_choice(fun, c("histogram", "density", "bar"))
+    assert_choice(fun, c("histogram", "density", "bar", "line"))
   } else {
     assert_function(fun, null.ok = TRUE)
   }
@@ -85,6 +80,9 @@ plot_tt <- function(x,
 
   } else if (identical(fun, "density")) {
     fun <- rep(list(tiny_density), length(data))
+
+  } else if (identical(fun, "line")) {
+    fun <- rep(list(tiny_line), length(data))
 
   } else if (identical(fun, "bar")) {
     for (idx in seq_along(data)) {
@@ -230,5 +228,15 @@ tiny_density <- function(d, color = "black", ...) {
 tiny_bar <- function(d, color = "black", xlim = 0:1, ...) {
   function() {
     graphics::barplot(d, horiz = TRUE, col = color, xlim = xlim)
+  }
+}
+
+
+tiny_line <- function(d, xlim = 0:1, color = "black", ...) {
+  function() {
+    if (!inherits(d, "data.frame") || !"x" %in% names(d) || !"y" %in% names(d)) {
+      stop("The data to plot a `line` must be a data frame with columns `x` and `y`.", call. = FALSE)
+    }
+    plot(d$x, d$y, type = "l", col = color, axes = FALSE, ann = FALSE, lwd = 50)
   }
 }