+++ R/style_bootstrap.R
@@ -16,10 +16,12 @@ style_bootstrap <- function(x,
                             fontsize = NULL,
                             width = NULL,
                             align = NULL,
+                            alignv = NULL,
                             line = NULL,
                             line_color = "black",
                             line_width = .1,
                             colspan = NULL,
+                            rowspan = NULL,
                             indent = 0,
                             bootstrap_class = NULL,
                             bootstrap_css = NULL,
@@ -80,6 +82,15 @@ style_bootstrap <- function(x,
     align_bootstrap <- align
   }
 
+  if (!is.null(alignv)) {
+    alignv_bootstrap <- switch(alignv,
+                              "t" = "top",
+                              "m" = "middle",
+                              "b" = "bottom")
+  } else {
+    alignv_bootstrap <- alignv
+  }
+
   if (!is.null(fontsize)) {
     fontsize_bootstrap <- sprintf("%sem", fontsize)
   } else {
@@ -94,6 +105,7 @@ style_bootstrap <- function(x,
   settings$bootstrap <- vectorize_bootstrap(settings$bootstrap, monospace, "font-family: monospace;")
   settings$bootstrap <- vectorize_bootstrap(settings$bootstrap, fontsize_bootstrap, "font-size: %s;")
   settings$bootstrap <- vectorize_bootstrap(settings$bootstrap, align_bootstrap, "text-align: %s;")
+  settings$bootstrap <- vectorize_bootstrap(settings$bootstrap, alignv_bootstrap, "vertical-align: %s;")
   settings$bootstrap <- vectorize_bootstrap(settings$bootstrap, color, "color: %s;")
   settings$bootstrap <- vectorize_bootstrap(settings$bootstrap, background, "background-color: %s;")
   settings$bootstrap <- vectorize_bootstrap(settings$bootstrap, width, "width: %s;")
@@ -113,6 +125,9 @@ style_bootstrap <- function(x,
   id <- sapply(unique(settings$bootstrap), function(k) get_id(stem = "tinytable_css_"))
   settings$id <- id[match(settings$bootstrap, names(id))]
 
+  if (is.null(rowspan)) rowspan <- 1
+  if (is.null(colspan)) colspan <- 1
+
   # CSS style for cell
   css_done <- NULL
   for (row in seq_len(nrow(settings))) {
@@ -120,6 +135,12 @@ style_bootstrap <- function(x,
     listener <- "window.addEventListener('load', function () { styleCell_%s(%s, %s, '%s') })"
     listener <- sprintf(listener, settings$id[row], settings$i[row], settings$j[row], settings$id[row])
     out <- bootstrap_setting(out, listener, component = "cell")
+
+    if (rowspan != 1 || colspan != 1) {
+      listener <- "window.addEventListener('load', function () { spanCell_%s(%s, %s, %s, %s) })"
+      listener <- sprintf(listener, settings$id[row], settings$i[row], settings$j[row], rowspan, colspan)
+      out <- bootstrap_setting(out, listener, component = "cell")
+    }
 
     # CSS styling
     css <- paste(bootstrap_css, settings$bootstrap[row], collapse = ";")
@@ -142,6 +163,7 @@ style_bootstrap <- function(x,
 
   # Changing function names to table ID to avoid conflict with other tables functions 
   out <- gsub("styleCell_\\w+\\(", paste0("styleCell_", meta(x, "id"), "("), out)
+  out <- gsub("spanCell_\\w+\\(", paste0("spanCell_", meta(x, "id"), "("), out)
 
   class(out) <- class(x)
   return(out)