+++ R/style_grid.R
@@ -6,6 +6,8 @@ style_grid <- function(x,
                        monospace = FALSE,
                        underline = FALSE,
                        strikeout = FALSE,
+                       rowspan = NULL,
+                       colspan = NULL,
                        ...) {
 
   if (meta(x, "output") != "markdown") return(x)
@@ -14,6 +16,20 @@ style_grid <- function(x,
 
   ival <- if (is.null(i)) seq_len(meta(x, "nrows")) else i
   jval <- if (is.null(j)) seq_len(meta(x, "ncols")) else j
+
+  # Unlike other formats, Markdown inserts `group_tt()` row labels after styling. This aligns the `i` index to the full columns.
+  gr <- meta(x, "lazy_group")
+  gr <- Filter(function(k) !is.null(k$i), gr)
+  # do not style spanning row labels
+  lab_idx <- drop(unlist(lapply(gr, function(k) k$i)))
+  lab_idx <- lab_idx + cumsum(rep(1, length(lab_idx))) - 1
+  ival <- setdiff(ival, lab_idx) 
+  for (g in gr) {
+    for (lab in g$i) {
+      ival[ival > lab - 1] <- ival[ival > lab - 1] - 1
+    }
+    lab_idx <- c(lab_idx, g$i)
+  }
 
   for (col in seq_along(out)) {
     out[[col]] <- as.character(out[[col]])
@@ -30,8 +46,23 @@ style_grid <- function(x,
       if (isTRUE(strikeout)) {
         out[row, col] <- sprintf("~~%s~~", out[row, col])
       }
+
+      if (!is.null(rowspan) || !is.null(colspan)) {
+        idx_row <- if (isTRUE(rowspan > 1)) row + seq_len(rowspan) - 1 else row
+        idx_col <- if (isTRUE(colspan > 1)) col + seq_len(colspan) - 1 else col
+        backup <- out[row, col]
+        for (w in idx_row) {
+          for (z in idx_col) {
+            if (z <= meta(x, "ncols")) {
+              out[w, z] <- ""
+            }
+          }
+        }
+        out[row, col] <- backup
+      }
     }
   }
+
 
   attr(out, "tinytable_meta") <- meta(x)
   class(out) <- class(x)