+++ R/style_tt.R
@@ -3,11 +3,14 @@
 #' @details
 #' This function applies styling to a table created by `tt()`. It allows customization of text style (bold, italic, monospace), text and background colors, font size, cell width, text alignment, column span, and indentation. The function also supports passing native instructions to LaTeX (tabularray) and HTML (bootstrap) formats.
 #'
-#' Warning: Markdown and Word formats are limited to these styles: italic, bold, strikeout. This is because there is no markdown syntax for the other options, and because we create Word documents by converting a markdown table to .docx via the Pandoc software.
+#' Note: Markdown and Word tables only support these styles: italic, bold, strikeout. Moreover, the `style_tt()` function cannot be used to style headers inserted by the `group_tt()` function; instead, you should style the headers directly in the header definition using markdown syntax: `group_tt(i = list("*italic header*" = 2))`. These limitations are due to the fact that there is no markdown syntax for the other options, and that we create Word documents by converting a markdown table to .docx via the Pandoc software.
 #'
 #' @param x A table object created by `tt()`.
 #' @param i Row indices where the styling should be applied. Can be a single value or a vector. `i=0` is the header, and negative values are higher level headers. If `colspan` is used, `i` must be of length 1.
-#' @param j Column indices where the styling should be applied. Can be a single value, a vector, or a Perl-style regular expression applied to column names of the original data frame. If `colspan` is used, `j` must be of length 1.
+#' @param j Column indices where the styling should be applied. Can be:
+#' + Integer vectors indicating column positions.
+#' + Character vector indicating column names.
+#' + A single string specifying a Perl-style regular expression used to match column names. 
 #' @param bold Logical; if `TRUE`, text is styled in bold.
 #' @param italic Logical; if `TRUE`, text is styled in italic.
 #' @param monospace Logical; if `TRUE`, text is styled in monospace font.
@@ -27,7 +30,9 @@
 #' @param width Width of column in em units. Can be `NULL` for default width.
 #' @param fontsize Integer Font size in pt units.
 #' @param align A single character or a string with a number of characters equal to the number of columns in `j`. Valid characters include 'c' (center), 'l' (left), or 'r' (right).
-#' @param colspan Number of columns a cell should span. Can only be used if both `i` and `j` are of length 1. Must be an integer greater than 1.
+#' @param alignv A single character specifying vertical alignment. Valid characters include 't' (top), 'm' (middle), 'b' (bottom).
+#' @param colspan Number of columns a cell should span. `i` and `j` must be of length 1.
+#' @param rowspan Number of rows a cell should span. `i` and `j` must be of length 1.
 #' @param indent Text indentation in em units. Positive values only.
 #' @param line String determines if solid lines (rules or borders) should be drawn around the cell, row, or column. 
 #' + "t": top
@@ -48,11 +53,20 @@
 #' @export
 #' @examples
 #' library(tinytable)
-#' x <- mtcars[1:5, 1:5]
+#' x <- mtcars[1:5, 1:6]
 #' tab <- tt(x)
-#' tab <- style_tt(tab, j = 1:5, align = "lcccr")
-#' tab <- style_tt(tab, i = 2:3, background = "black", color = "orange", bold = TRUE)
+#' 
+#' # Alignment
+#' style_tt(tab, j = 1:5, align = "lcccr")
+#' style_tt(tab, i = 2:3, background = "black", color = "orange", bold = TRUE)
 #' tab
+#' 
+#' # column selection with `j``
+#' x <- mtcars[1:5, 1:6]
+#' tab <- tt(x)
+#' style_tt(tab, j = 5:6, background = "pink")
+#' style_tt(tab, j = "drat|wt", background = "pink")
+#' style_tt(tab, j = c("drat", "wt"), background = "pink")
 #'
 style_tt <- function (x,
                       i = NULL,
@@ -67,7 +81,9 @@ style_tt <- function (x,
                       fontsize = NULL,
                       width = NULL,
                       align = NULL,
+                      alignv = NULL,
                       colspan = NULL,
+                      rowspan = NULL,
                       indent = 0,
                       line = NULL,
                       line_color = "black",
@@ -95,7 +111,9 @@ style_tt <- function (x,
               fontsize = fontsize,
               width = width,
               align = align,
+              alignv = alignv,
               colspan = colspan,
+              rowspan = rowspan,
               indent = indent,
               line = line,
               line_color = line_color,
@@ -130,7 +148,9 @@ style_tt_lazy <- function (x,
                            fontsize,
                            width,
                            align,
+                           alignv,
                            colspan,
+                           rowspan,
                            indent,
                            line,
                            line_color,
@@ -147,12 +167,12 @@ style_tt_lazy <- function (x,
 
   out <- x
 
-  # j is a regular expression
-  if (is.character(j) && !is.null(meta(x, "colnames"))) {
-    j <- grep(j, meta(x, "colnames"), perl = TRUE)
-  }
+  j <- sanitize_j(j, x)
 
-  # align can be "c" or "clrrlc"takes many possible values
+  # alignv can only be a single character for now
+  assert_choice(alignv, c("t", "m", "b"), null.ok = TRUE)
+
+  # align can be "c" or "clrrlc" takes many possible values
   assert_string(align, null.ok = TRUE)
 
   if (!is.null(align)) {
@@ -185,7 +205,8 @@ style_tt_lazy <- function (x,
 
   assert_style_tt(
     x = out, i = i, j = j, bold = bold, italic = italic, monospace = monospace, underline = underline, strikeout = strikeout,
-    color = color, background = background, fontsize = fontsize, width = width, align = align, colspan = colspan, indent = indent,
+    color = color, background = background, fontsize = fontsize, width = width, align = align, alignv = alignv, 
+    colspan = colspan, rowspan = rowspan, indent = indent,
     line = line, line_color = line_color, line_width = line_width,
     tabularray_inner = tabularray_inner, tabularray_outer = tabularray_outer, bootstrap_css = bootstrap_css,
     bootstrap_css_rule = bootstrap_css_rule, bootstrap_class = bootstrap_class)
@@ -196,25 +217,25 @@ style_tt_lazy <- function (x,
 
   out <- style_tabularray(
     x = out, i = i, j = j, bold = bold, italic = italic, monospace = monospace, underline = underline, strikeout = strikeout,
-    color = color, background = background, fontsize = fontsize, width = width, align = align, colspan = colspan, indent = indent,
+    color = color, background = background, fontsize = fontsize, width = width, align = align, alignv = alignv, colspan = colspan, rowspan = rowspan, indent = indent,
     tabularray_inner = tabularray_inner, tabularray_outer = tabularray_outer,
     line = line, line_color = line_color, line_width = line_width)
 
   out <- style_bootstrap(
     x = out, i = i, j = j, bold = bold, italic = italic, monospace = monospace, underline = underline, strikeout = strikeout,
-    color = color, background = background, fontsize = fontsize, width = width, align = align, colspan = colspan, indent = indent,
+    color = color, background = background, fontsize = fontsize, width = width, align = align, alignv = alignv, colspan = colspan, rowspan = rowspan, indent = indent,
     bootstrap_css = bootstrap_css, bootstrap_css_rule = bootstrap_css_rule, bootstrap_class = bootstrap_class,
     line = line, line_color = line_color, line_width = line_width)
 
   out <- style_grid(
     x = out, i = i, j = j, bold = bold, italic = italic, monospace = monospace, underline = underline, strikeout = strikeout,
-    color = color, background = background, fontsize = fontsize, width = width, align = align, colspan = colspan, indent = indent,
+    color = color, background = background, fontsize = fontsize, width = width, align = align, alignv = alignv, colspan = colspan, rowspan = rowspan, indent = indent,
     bootstrap_css = bootstrap_css, bootstrap_css_rule = bootstrap_css_rule,
     line = line, line_color = line_color, line_width = line_width)
 
   out <- style_typst(
     x = out, i = i, j = j, bold = bold, italic = italic, monospace = monospace, underline = underline, strikeout = strikeout,
-    color = color, background = background, fontsize = fontsize, width = width, align = align, colspan = colspan, indent = indent,
+    color = color, background = background, fontsize = fontsize, width = width, align = align, alignv = alignv, colspan = colspan, rowspan = rowspan, indent = indent,
     bootstrap_css = bootstrap_css, bootstrap_css_rule = bootstrap_css_rule,
     line = line, line_color = line_color, line_width = line_width)
 
@@ -235,7 +256,9 @@ assert_style_tt <- function (x,
                              fontsize,
                              width,
                              align,
+                             alignv,
                              colspan,
+                             rowspan,
                              indent,
                              line,
                              line_color,
@@ -253,7 +276,8 @@ assert_style_tt <- function (x,
     stop(msg, call. = FALSE)
   }
 
-  assert_integerish(colspan, len = 1, lower = 1, null.ok = TRUE)
+  assert_integerish(colspan, len = 1, lower = 2, null.ok = TRUE)
+  assert_integerish(rowspan, len = 1, lower = 2, null.ok = TRUE)
   assert_numeric(width, len = 1, lower = 0, null.ok = TRUE)
   assert_numeric(indent, len = 1, lower = 0)
   assert_character(background, null.ok = TRUE)